path smoothing implemented
- had some issues with mem leaks but got it figured out...eventually
- path smoothing works with A*, Dijkstra, & Depth 1st by toggle of #define SMOOTH_PATH in PathSmooth.h

+++ GameAI/pathfinding/game/component steering/ArriveSteering.h
@@ -10,7 +10,7 @@ class ArriveSteering : public Steering
 		ArriveSteering(const UnitID& ownerID, const Vector2D& targetLoc, const UnitID& targetID = INVALID_UNIT_ID, bool shouldFlee = false);
 		virtual Steering* getSteering(); //overrides parent function
 
-		int getTargetRadius() { return mTARGET_RADIUS;}
+		float getTargetRadius() { return mTARGET_RADIUS;}
 
 	protected:
